Add dummy app steps (adapted to the Rails engine context)
Add Dummier
See https://github.com/citrus/dummier

I'm not sure why, but the `rails new` generator can't be run
into a Rails engine context. And that prevented the dummy app generation.

On the other side, running the Rails application generator from Dummier
works pretty well. Again, I'm not sure why, but it does work.
See features/successful_dummy_features.feature for live examples.

Drawback: the Cucumber features were alreay slow when they involved
the generation of a dummy app. Dummier adds it's own slowness to the drama.
Personnally, I think the regression tests and the documentation we get
are worth the time needed to generate them, but the performance is now
an issue which should be tackeled as soon as possible.